Added creation of input mail from python message
That way we don't need extra logic for the welcome
mail, we just have to read the file and send the contents
to the input mail parser and that's it.

Also moved the logic of adding a welcome mail to the mailboxes
because it has knowledge of mailbox methods anyways.