[test] leap_mailstore handles unknow uuid as empty
Given a unknown mailbox uuid, we should return an empty string as name.
This will show the mail as in 'all mails' instead of a KeyError.
This scenario is supposed to happen only if the mail comes before the
mailbox, which is our current unicorn. Anyway, when the missing mailbox
gets created it will read correctly or let the see under all mails
instead of breaking the account.
--Issue #444