Removed search engine from constant in Indexer
The search engine was being passed as a constant to
the Mail Indexer, that constrained the user agent to
one user (because the search engine is user specific).

I added the search engine as a parameter on the Mail Indexer
initialization so that we can have the Indexer working for
each user.