Issue #292: put labels on user pub key and email
Since we need always think about our defined personas, it's important to
describe properly the information we put on user interface. This aspect
is more important when we need to show some non trivial information as
the public key fingerprint.

So, this change adds label for both info on user "settings" box, email
and public user fingerprint.