Fixed mockito conflict with zope
I bumped into this problem where trial would try to clean up after
running the SiteTest and run into an error
'RememberedInvocation has no attribute "_implied"'

That happened because mockito is strict with the set of functions
it's mock accepts and _implied doesn't exist. It didn't really make
sense in this test context so I've adapted the test to use MagicMock
instead and now the test pass without problems.